requirements.txt
# Core dependencies
PyGithub==2.1.1
notion-client==2.2.1
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# ML dependencies (see ml/requirements.txt for full ML stack)
scikit-learn==1.3.2

# CLI utilities
rich==13.7.0
click==8.1.7

# Security
cryptography==41.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
